requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2025.4.26
charset-normalizer==3.4.2
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.58.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.0
omegaconf==2.3.0
onnxruntime==1.22.0
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
progress==1.6
protobuf==6.31.1
pyclipper==1.3.0.post6
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
rapidocr==3.0.0
rapidocr-onnxruntime==1.4.4
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.1
six==1.17.0
sympy==1.14.0
tifffile==2025.5.10
tqdm==4.67.1
urllib3==2.4.0
